+++ src/pages/messages/TasksPage.tsx
@@ -142,10 +142,12 @@ const TasksPage = () => {
     .filter(Boolean)
     .join(' AND ');
 
+  const ticketQuery = `${ticketQueryString}&viewingScope=${organizationFilter.join(',')}`;
+
   const ticketsQuery = useQuery({
     enabled: isOnTicketsPage,
-    queryKey: ['tickets', rowsPerPage, page, ticketQueryString],
-    queryFn: () => fetchTickets(rowsPerPage, (page - 1) * rowsPerPage, ticketQueryString),
+    queryKey: ['tickets', rowsPerPage, page, ticketQuery],
+    queryFn: () => fetchTickets(rowsPerPage, (page - 1) * rowsPerPage, ticketQuery),
     meta: { errorMessage: t('feedback.error.get_messages') },
   });
